First Stab at CreateAttributeAssembledViews.js [Aergo-529]
This is a basic view that joins the posit and annex tables of an attribute together. The Reliable column is current commented out awaiting support in the underlying attribute annex table. An additional TODO comment notes that the unique clustered index on a view is an mssql feature absent in postgresql that could warrant further research.